requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2
lightgbm>=4.0
statsmodels>=0.14
arch>=6.3
torch>=2.1
matplotlib>=3.7
pyyaml>=6.0
ccxt>=4.0
joblib>=1.3
binance-bulk-downloader>=1.1.0.1
